Add support for 32U4
Added support for AVR 32U4. Due to an issue how COM (with 32U4) is handled in windows, not allowing the Arduino to send feedback while trying to program, it's only working using Linux. Arduino libraries are being used for init of USB and also for some LED flashing added for finding where the sketch stops working using windows.